pyproject.toml
[project]
name = "clpga-demo"
version = "0.1.0"
description = "Track golf balls in video using SAM3 and output portrait-cropped video"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ultralytics>=8.4.23",
    "opencv-python-headless>=4.8",
    "scipy>=1.11",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "clip",
    "timm>=1.0.25",
    "huggingface-hub>=0.34.0,<1.0",
    "google-genai>=1.71.0",
]

[project.scripts]
clpga-demo = "clpga_demo.__main__:main"

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
clip = { git = "https://github.com/ultralytics/CLIP.git" }

[dependency-groups]
dev = [
    "pytest>=8.0",
]
